134.2. Helper Functions in Function Components(Section 12: Understanding Hooks in React)
Let me show you how we would approach adding in some kind of helper function or helper method to a function

component.

I'm gonna delete that little example.

Then at the very top of accordion, I'm going to define a new helper function and assign it to a variable

called on Idol Click.

I'm going to expect that this is going to receive the index of the item that was just clicked on and

then inside there will do a console log as well of title clicked and then print out the index.

So I now want to call this function right here.

Anytime a user clicks on that div.

So now I'm still going to leave in this aero function right here and I'm going to replace that console

log.

And call on title click and pass through the index, though, on Idol, click and pass through the index

like so.

Why do we still have the aero function?

Well, remember, if we just write out something like this.

Then on Title Click is going to be invoked the instant that our list of items is rendered.

As a matter of fact, we can even verify that if we save the file as it is right now.

Flip back over and the instant we refresh the page, we see title click zero one into.

We do not want to run that callback function the instant our company is rendered.

Instead, we want to call it at some point in time the future.

And to do that, we have to wrap it with an arrow function.

Now, whenever that era function is executed, we run on title click.

And the reason I'm making such a big deal about this entire helper method right here is that it's gonna

be a little bit nasty and your components are gonna end up being a little cluttered.

When you write out all these helper functions inside the body of a function component, you don't get

that kind of nice separation of code that we used to have back inside of a class based component.

So this is just something I want you to be aware of a little bit harder to write, a very nicely organized

component.

When you are using a function component, that's just the reality.

All right.

Let's save this with the arrow function in there.

And then do another quick test.

Well, I'll click on these titles, and yet I still see those console logs.

OK.

So this looks pretty good.

Now, as I mentioned, we need to make use of the Hooke's state system to somehow keep track of which

of these elements should be expanded and collapsed.

Let's take a pause right here and then start to understand the state system with hooks in the next video.