147.3. Testing Execution(Section 12: Understanding Hooks in React)
Now, the last Nerio is that we have an array with something inside of it.

In this case, we'll put in term.

Now, term being put inside of here is going to give us a little bit misleading effect when we put in

term or anything inside that array.

Remember, that means you want to run at the initial render and after every re render, if that data

has changed.

Well, in this case, the only way that we can rebrand or are component is if term changes.

So this right now with turn right here is going to appear to have the same effect as no array whatsoever

because it's going to run this use effect function any time term changes.

We'll try it out any ways, even though this will be just a little bit misleading.

So if I type inside of your yep, I see the same thing.

OK, so that clarifies when this erro function gets executed.

So remember, it's all about that second argument.

It is rather rare for us to not provide any second array or argument or that array at all.

Much more frequently.

You're going to see either an empty array or when we want to run code during the initial render, or

we are going to see an array with something inside of it like.

So that is way more common than seeing nothing inside there at all.

Now, last thing I want to mention very quickly, this array can have multiple elements inside of it

that we had some other piece of state inside of here, let's say like term two or something.

We can't have term to side there and now use effect is going to run.

If either of those elements change between renders, so only one has to change and the aero function

will be executed.

Again.

All right.

So we now have a pretty good idea.

I'm going to take out that term to piece of state.

I think we've now got a pretty good idea.

About what use effect is really doing and how we can customize it.

Now we're going to make use of use effect to take that term after it changes and make a request out

to the Wikipedia API.

Well, we'll take you to that request in just a moment.