148.1 Async Code in useEffect(Section 12: Understanding Hooks in React)
Well, I think we know enough about use effect.

So let's now try to make a request from inside of that aero function to make requests.

We're going to need to install Axios at my terminal.

I'm going to make sure I'm inside of my widgets project directory.

And then I will do in NPM install Texaco's.

All right, I'll let that do its thing while it's running.

I got to flip back over to my file at the top.

I'm going to import Axios from Axios.

Now, before we make our request, we need to figure out how we should configure use effect.

Let's clean up some of these council logs.

And I'm going to remove that second argument from use effect for right now before we really decide how

we should configure it.

I want you to remember the goal of this widget that we're putting together.

We said looking at that mockup.

Right here, we said that we want to do a search of the Wikipedia API every single time a user presses

a key, the immediately after user presses a key.

We want to do a search.

That means we're gonna make a lot of requests very quickly.

But as I mentioned, we will try to fix that up later on.

So how should we configure use effect to make sure that we run a request every single time?

Our component re renders.

And that term piece of state changes.

Well, to do so.

We could find use fact right here and we would put in a second argument.

It would be an array and we would write in their term.

So this means whenever we re render our components and the term has changed, Ron, that aero function

right there.

It also means that we're going to immediately run the ROV function when our components first rendered

as well.

So I would say that that pretty well meets the requirements of our application.

We're going to have the opportunity to run some code right here anytime term changes.

So I think that we can leave use effect as it is right now.

And just go ahead and make our requests inside of you're using axios.

All right.

So the depend this array right here is what we want.

That looks good.

We need to make the request inside of here.

Now, usually what you and I would want to do, based upon everything we've learned throughout this

course so far, we would probably want to mark this function as async and then do a request inside of

here with a wait axios and then some your L.

Well, that would be great if we could do that, but unfortunately, we can't.

There's a little requirement around use effect.

We are not allowed to mark the function that we are passing into use effect.

So this area function right here, we can not mark that as async and use any await keywords directly

inside that function.

If we tried to save this code right now and flip back over to our browser.

You can open up your console, and if you scroll around a little bit, you're going to see a really

long error message like this right here.

So you can read this error message if you wish.

It is essentially saying that we are not allowed to mark the arrow function that we are passing into

to use effect as async.

There are a couple of different ways we can work around this.