148.2 Async Code in useEffect(Section 12: Understanding Hooks in React)
So what I have right here not allowed.

We have to use some alternative solution.

There are in total, three different ways that we can still make a request inside of use effect.

We can absolutely make a request.

Don't get me wrong.

The only requirement is we cannot mark this thing as being a sink.

That's all.

So to get around that.

Let me show you the three possible solutions.

The first solution that we can use is going to be essentially what is recommended in the air message

right here.

Inside of use effect, we can make a little temporary helper function.

We could call it something like.

Search articles or even maybe just search Wikipedia.

Just search.

We'll leave it really simple and straightforward.

Now we can define a function inside of use effect and mark that one as being async.

We can then use the await keyword inside of your as much as we wish.

So we could call Accio Scott, get make a request and do whatever else.

And then immediately after.

We define that function.

We can call it this right here is allowed.

And if you follow this error message, it's kind of the recommended approach.

We could save this flip back over and now we will see an error around the request because I don't have

a real you are l inside there.

You'll notice that we are no longer getting an air from react itself.

That's method number one that we can use to still make use the async 08 syntax inside of async function,

ask me.

I use the effect function.