148.3 Async Code in useEffect(Section 12: Understanding Hooks in React)
Method number two is going to very similar, but you'll notice that with this approach right here,

we are making a temporary variable and then calling it meta after.

So in solution method number two, you're going to see this somewhat frequently in the wild on professional

projects.

We can remove that temporary variable altogether or move the cont search and the search invocation down

there.

If we just write this out, this isn't going to do anything for us.

We've declared a function, but it doesn't get invoked.

So to invoke it.

We could wrap this with a set of parentheses.

And then after the closing parentheses right here, we'll put in another set of parentheses.

And the Syntex looks all a bit strange, but it is 100 percent valid JavaScript, this defines a function.

And then immediately invokes it.

So it's equivalent a hundred percent equivalent to what we just had a moment ago.

It just removes that extra variable declaration.

That's all.

If you're thinking that this would be superior from a performance perspective because we're not creating

the additional variable.

Don't get too excited.

The actual performance overhead of creating a variable is absolutely miniscule.

So you would only want to take this approach right here.

If you think the code is easier to read, do not do this because you think there is any kind of performance

benefit or anything like that.