148.4 Async Code in useEffect(Section 12: Understanding Hooks in React)
Now, the third possible approach.

So this is number two, their possible approach is to just revert back to using normal promises.

Whenever you make a request to Axios.

It gives us back a promise we could chain on a dot van statement.

And this erro function right here will be invoked with the response that we get back in that API.

And so we could do a console log of response DOT data and they would be all the information we are looking

for.

So those are the three possible alternatives.

Either declare the helper function, declare a helper function and Amela invoke it, or just make use

of promises.