148.5 Async Code in useEffect(Section 12: Understanding Hooks in React)
React itself, based upon that air message you just saw, pretty much recommends solution number one,

which is this right here.

So declare the helper function and then just call it yourself immediately.

I would say that in the community, there's no really strong direction one way or another.

Certainly I think promises are the least often used, but sometimes they actually are the easiest to

use.

To be totally honest with you, in this course, we're going to generally trend toward using this method

right here.

But sometimes we will swap out to using a promise if the syntax ends up being just a little bit easier

to write out.

OK.

So that's how we can the three ways that we can make a request inside of use effect while still using

some async await syntax or a promise.

Can you really have to remember here?

Don't try to mark that function as async.

Otherwise, you're going to get an error.

Well, let's take a pause right here and then put in the real you l for a request in just a moment.