150.1. Default Search Terms(Section 12: Understanding Hooks in React)
Let's add in a new piece of state to our application, though, in the search component.

I'm going to initialize a new piece of state at the very top.

I'm going to call this new piece of state results because it is the results of my search operation.

We'll get results and set results.

We then need to think of a default value for this piece of states.

It is eventually going to be an array of objects.

I think that when we first render our component to the screen, it would make sense to use a default

value here of an empty array.

Just to say we do not have any search results right now.

Then inside this search function, we're going to take the data out of the response we get.

We have concerns, data is the request.

And then right after we make the request.

Make sure that you're putting this still inside of this search async function.

We're going to use data to update our results.

Piece of state.

So we will put in set results.

And provide data.

That will cause our component to automatically re render.

So we can then access the results array down inside of our J as X block and render out our list.

Right now, let's just do a console log of results and make sure that we're getting the correct data.

All right, back over here.

Well, it looks like this first request did not quite go as expected.

If I go to my console, you'll notice I get an empty array one time slots from when our component first

renders.

And then on this next console log, we get back in air and it says, as our search parameter must be

set.

Why is that?

Well, when our component is rendered for the very first time, term is an empty string.

If we provide an empty string to the Wikipedia API, they are saying, hey, you have to provide us

and s our search.