150.2 Default Search Terms(Section 12: Understanding Hooks in React)
You can't search for empty string.

We have two options available to us.

We can either put in a default search term that will be displayed to the user.

For example, we could put in programming.

Right here.

If we do that, then when our components first rendered, we will initiate a search and we will search

for the term programming, that is option number one.

Let's try to see this and then see what happens.

You'll notice now that we've got an initial value of our state of programming and that shows up inside

of our inputs.

And then we could see the response data.

Right here, we can expand query.

And those are some computer programming articles.