150.3 Default Search Terms(Section 12: Understanding Hooks in React)
If we do not want to provide an initial value for our user, but we don't want to the request to air

out or like that, we could revert you state right here and give it an empty string as the default value.

And then inside of use effect.

When we call search right here, rather than immediately calling search, we can make sure.

That term is defined.

But now, if term is an empty string.

For me, a term it will, if term it is an empty string, will not search if term has some characters

inside of it.

We will do a search though, if we take this approach.

We will not have a default search term and we will not make an initial request until the user starts

to type something in like so.

So it's really up to you.