150.4 Default Search Terms(Section 12: Understanding Hooks in React)
Maybe rather than taking the entire data property out of that response and using that to update our

results piece of state, maybe we should kind of crawl through there and just get this search array

of results.

And set that on our state instead.

Before that, when we sat results with data, we could change it to data dot the data.

Is this entire object to be data?

Dot query.

Dot search.

Get let's save that again.

And now we can see that during our initial render, we have no results.

The request is then completed.

We update our states.

We do a console log of the results piece of state, and we can see that we've got the array of objects.

All right.

What's the.

It's looks pretty good.

Then the last thing I'll do is remove that console log.

And then finally, we need to map over that results, Ray, and build out our actual list of elements

to display inside this giant x block.

Let's take care of that list building in the next video.