151.2 List Building!(Section 12: Understanding Hooks in React)
Let's first take care of fixing up that key property and then we'll figure out how to remove that ugly

looking H.T. inside there.

After that, to remove the key property.

You'll notice that in the response we get back.

Every result we have has a page I.D. and so we can use this page I.D. as the key property because as

usual, it is probably consistent and unique.

We're going to find my list on that d'Hiv.

I'll put in the key property as a result.

That age I.D. and I want you to take a really close look here.

I.D. is lowercase.

We need to put in page I.D. with a lowercase I.D., though, page I.D..

Make sure it's lowercase.

When we're safe.

Well, it looks like that era has gone away.

Now we just need to figure out all this H.T. mail stuff.

Let's figure out how to remove all that H.T. mail in just a moment.